% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{aggregate_by_hierarchy}
\alias{aggregate_by_hierarchy}
\title{Aggregate state fractions by hierarchy}
\usage{
aggregate_by_hierarchy(state.mat, hierarchy)
}
\arguments{
\item{state.mat}{Samples x states matrix.}

\item{hierarchy}{Hierarchy table.}
}
\value{
list with subtype and major matrices.
}
\description{
Aggregate state fractions by hierarchy
}
