% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_experiment1.R
\name{build_bayesprism_panel}
\alias{build_bayesprism_panel}
\title{Build BayesPrism-panel output from state/subtype matrices}
\usage{
build_bayesprism_panel(state.mat)
}
\arguments{
\item{state.mat}{Sample x state matrix.}
}
\value{
Sample x panel matrix.
}
\description{
This helper maps our output to a common BayesPrism-style panel: CD4Tn,
CD8Tn, FOB, ILC3, TAM, Treg, cDC1, cDC2, CMonocyte, NMonocyte, pDC,
and Thelper.
}
