% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_reference.R
\name{build_merge_map}
\alias{build_merge_map}
\title{Build merge map}
\usage{
build_merge_map(original.cells, merge.groups)
}
\arguments{
\item{original.cells}{Original cell-type names.}

\item{merge.groups}{Named merge groups.}
}
\value{
Named character vector mapping original cells to merged cells.
}
\description{
Build merge map
}
