% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchy.R
\name{refine_monocyte_subtypes}
\alias{refine_monocyte_subtypes}
\title{Refine monocyte subtype labels in a hierarchy table}
\usage{
refine_monocyte_subtypes(hierarchy)
}
\arguments{
\item{hierarchy}{data.frame with state_name, subtype, major_lineage.}
}
\value{
Modified hierarchy.
}
\description{
This splits cMo/intMo/ncMo into independent subtypes under Myeloid.
}
