% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_deconvolution_outputs}
\alias{write_deconvolution_outputs}
\title{Write deconvolution outputs}
\usage{
write_deconvolution_outputs(result, out.dir)
}
\arguments{
\item{result}{Result from deconvolve_bulk_matrix().}

\item{out.dir}{Output directory.}
}
\description{
Write deconvolution outputs
}
